requirements.txt
arrow >= 1.2.3
Flask>=2.2.2

# google.auth >= 2.11.0
# google-cloud-documentai == 1.2.0
google-cloud-documentai == 2.18.0
google-cloud-firestore >= 2.11.1
# google-cloud-pubsub >= 2.13.6
# google-cloud-storage>=1.42.0
google-cloud-secret-manager >= 2.16.3

gunicorn==20.1.0

python-dotenv >= 0.21.0
pytest-flask
requests==2.31.0
# structlog==21.5.0

spacy